cleanup and add filter and sort state
